dom/serviceworkers/test/test_serviceworker_header.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test that service worker scripts are fetched with a Service-Worker: script header</title>
  <script type="text/javascript" src="http://mochi.test:8888/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="http://mochi.test:8888/tests/SimpleTest/test.css" />
  <base href="https://mozilla.org/">
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  function runTest() {
    navigator.serviceWorker.register("http://mochi.test:8888/tests/dom/workers/test/serviceworkers/header_checker.sjs")
      .then(reg => {
        ok(true, "Register should succeed");
        reg.unregister().then(() => SimpleTest.finish());
      }, err => {
        ok(false, "Register should not fail");
        SimpleTest.finish();
      });
  }

  SimpleTest.waitForExplicitFinish();
  onload = function() {
    SpecialPowers.pushPrefEnv({"set": [
      ["dom.serviceWorkers.exemptFromPerDomainMax", true],
      ["dom.serviceWorkers.testing.enabled", true],
      ["dom.serviceWorkers.enabled", true],
    ]}, runTest);
  };
</script>
</body>
</html>
